requirements.txt
chalkpy>=1.11.8
pydantic
cattrs
requests